追加実装

+++ app/Http/Controllers/PostController.php
@@ -61,7 +61,11 @@ class PostController extends Controller
     function destroy($id)
     {
         $post = Post::find($id);
+        if(Auth::id() != $post->user_id) {
+            return abort(404);
+        }
         $post->delete();
         return redirect()->route('posts.index');
+
     }
 }
